4565: Fixed remaining tests
