+++ assets/tests/admin/data-fixtures.js
@@ -1007,6 +1007,642 @@ const screensListJson = {
   ],
 };
 
+const templatesListJson = {
+  "@id": "/v2/templates",
+  "hydra:member": [
+    {
+      "@type": "Template",
+      "@id": "/v2/templates/00XZXR5XDH0D1M16K10NYQ0A55",
+      title: "Est totam provident sunt.",
+      description:
+        "Tempora qui minus officia quis consequuntur voluptates. Quasi minima eveniet repudiandae laborum dolor quasi totam qui. Iusto enim inventore molestias amet aut.",
+      created: "2002-08-30T14:14:07+02:00",
+      modified: "2021-12-09T12:01:33+01:00",
+      modifiedBy: "",
+      createdBy: "",
+      resources: {
+        admin:
+          "http://www.harber.com/atque-inventore-consequatur-mollitia-ducimus-veritatis-doloribus-ad",
+        schema: "http://www.kulas.net/quia-unde-quos-error-modi-saepe",
+        component: "http://keebler.com/",
+        assets: {
+          type: "css",
+          url: "https://www.borer.biz/voluptas-blanditiis-et-quo-aut-culpa-reiciendis-dolorum",
+        },
+        options: {
+          fade: true,
+        },
+        content: {
+          text: "Accusantium exercitationem animi qui provident ipsa distinctio.",
+        },
+      },
+    },
+    {
+      "@type": "Template",
+      "@id": "/v2/templates/016MHSNKCH1PQW1VY615JC19Y3",
+      title: "Ut exercitationem est quia ad quas.",
+      description:
+        "Laborum quod ut ducimus suscipit quia nostrum. Saepe ex voluptas aut. Sit numquam vel est sunt. Cupiditate excepturi non saepe in voluptatem vel rem quaerat. Magni aut eaque vel deleniti.",
+      created: "2012-01-28T09:17:22+01:00",
+      modified: "2021-12-09T12:01:33+01:00",
+      modifiedBy: "",
+      createdBy: "",
+      resources: {
+        admin:
+          "https://vandervort.com/sapiente-quo-est-rerum-nihil-sint-placeat-ipsa-id.html",
+        schema:
+          "http://www.gulgowski.com/debitis-voluptatem-earum-sed-totam-aut-impedit-facere",
+        component:
+          "https://lehner.com/officia-ducimus-ea-beatae-eum-amet-provident-sint.html",
+        assets: {
+          type: "css",
+          url: "http://www.emmerich.net/aliquam-excepturi-id-et-ab-voluptate.html",
+        },
+        options: {
+          fade: true,
+        },
+        content: {
+          text: "Ut qui assumenda ex vel quod dolorem perspiciatis eos quis.",
+        },
+      },
+    },
+    {
+      "@type": "Template",
+      "@id": "/v2/templates/017X81AEJV2GJE0NC51KKK0EK8",
+      title: "Delectus magnam repudiandae molestiae et a.",
+      description:
+        "Consequuntur est ut commodi sed. Fugiat repellat harum assumenda sed illo voluptatem nobis fugit. At vero consequatur ut dignissimos. Et inventore ipsam ullam ullam dolor debitis quo saepe.",
+      created: "2013-06-17T03:02:15+02:00",
+      modified: "2021-12-09T12:01:33+01:00",
+      modifiedBy: "",
+      createdBy: "",
+      resources: {
+        admin: "http://www.ratke.com/libero-provident-nihil-minus-alias",
+        schema: "http://murazik.biz/",
+        component:
+          "http://kuhlman.biz/inventore-iure-tempora-perspiciatis-repudiandae-numquam-veniam-sequi-dolorem.html",
+        assets: {
+          type: "css",
+          url: "http://dibbert.biz/eius-et-non-autem",
+        },
+        options: {
+          fade: true,
+        },
+        content: {
+          text: "Praesentium at porro aut corporis quis in quia asperiores sed sit.",
+        },
+      },
+    },
+  ],
+  "hydra:totalItems": 12,
+};
+
+const slidesListJson = {
+  "@id": "/v2/slides",
+  "hydra:member": [
+    {
+      "@id": "/v2/slides/0086TQQC671WHA1S150MMF1Q3T",
+      title: "Adipisci vero quia.",
+      description:
+        "Dolores porro ex sed consectetur dolorem aspernatur. Recusandae voluptatem non a aut. Tenetur optio aut fugit reprehenderit. Non dolorum temporibus possimus iure aut quas.",
+      created: "1978-12-11T09:47:36.000Z",
+      modified: "2021-12-09T12:01:33.000Z",
+      modifiedBy: "",
+      createdBy: "",
+      templateInfo: {
+        "@id": "/v2/templates/00MWCNKC4P0X5C0AT70E741E2V",
+        options: [],
+      },
+      theme: "",
+      onPlaylists: ["/v2/playlists/00S7ZQK8Y90R351YES1DJN0RKR"],
+      duration: 70592,
+      published: {
+        from: null,
+        to: "1989-08-28T18:14:52.000Z",
+      },
+      media: ["/v2/media/00W2E6VC0V22QK1WYH0CMP151C"],
+    },
+    {
+      "@id": "/v2/slides/00BEFT32281WT51SNJ0JXA11AV",
+      title: "Alias et id consequatur.",
+      description:
+        "Beatae reiciendis provident eos est totam repudiandae molestiae. Qui itaque officiis quibusdam. A nisi minus dolorum excepturi. Atque molestiae non velit ipsum consequatur nisi.",
+      created: "1982-06-21T15:09:47.000Z",
+      modified: "2021-12-09T12:01:33.000Z",
+      modifiedBy: "",
+      createdBy: "",
+      templateInfo: {
+        "@id": "/v2/templates/00PXXK8ND01SRF18YH1WN004VA",
+        options: [],
+      },
+      theme: "/v2/themes/01FPFH3WX93S4575W6Q9T8K0YB",
+      onPlaylists: [],
+      duration: 40787,
+      published: {
+        from: "2021-03-12T18:12:47.000Z",
+        to: "2021-05-20T08:42:21.000Z",
+      },
+      media: [
+        "/v2/media/0010X8D6JJ03G50T1J1FCW1XH6",
+        "/v2/media/0043AB9VWE08071MNB08Q20JSG",
+        "/v2/media/009RGQK5AD19K21CZE11HC1DPX",
+        "/v2/media/00HM3QQ61107C41D7M0W2B1CEK",
+        "/v2/media/01BTDHQH3Z06N40M9V0TJ41YA8",
+      ],
+    },
+    {
+      "@id": "/v2/slides/00BCX948J60YCR1TC009A31MPQ",
+      title: "Amet impedit eum quia quo.",
+      description:
+        "Quis nostrum voluptatum quam aut sint enim. Ut sed occaecati deserunt accusantium. Cupiditate et voluptatum ex quidem. Nobis rerum consequatur fugiat occaecati voluptas et.",
+      created: "1982-06-02T00:11:19.000Z",
+      modified: "2021-12-09T12:01:33.000Z",
+      modifiedBy: "",
+      createdBy: "",
+      templateInfo: {
+        "@id": "/v2/templates/000YR9PMQC0GMC1TP90V9N07WX",
+        options: [],
+      },
+      theme: "",
+      onPlaylists: [],
+      duration: 4902,
+      published: {
+        from: null,
+        to: null,
+      },
+      media: [
+        "/v2/media/00007SBZ470CJ60C7J06H508R8",
+        "/v2/media/004MDCEC451GPJ1DW80D1Z026F",
+      ],
+    },
+    {
+      "@id": "/v2/slides/000AGMKXFZ0QWX0V5013R80S3S",
+      title: "Aperiam maxime autem.",
+      description:
+        "Ea laboriosam rerum voluptatem. Quos odit veniam cum. Quia non expedita non facere id eum nesciunt. Fugit dolorum ipsa aspernatur.",
+      created: "1970-05-11T17:45:13.000Z",
+      modified: "2021-12-09T12:01:33.000Z",
+      modifiedBy: "",
+      createdBy: "",
+      templateInfo: {
+        "@id": "/v2/templates/018A0WH44D0X1N176Z16CC0HJP",
+        options: [],
+      },
+      theme: "",
+      onPlaylists: [],
+      duration: 85133,
+      published: {
+        from: "2021-08-13T23:56:36.000Z",
+        to: null,
+      },
+      media: [
+        "/v2/media/00CGESC8E10CZA11EV2AYR1KWC",
+        "/v2/media/00HM3QQ61107C41D7M0W2B1CEK",
+        "/v2/media/01DCA32QJY1BH600BV2H140JDK",
+      ],
+    },
+    {
+      "@id": "/v2/slides/014FR06WMH05WP17B71Q6Z1K7R",
+      title: "Architecto dolores facilis et.",
+      description:
+        "Aspernatur doloremque et perferendis dolorum hic. Ab vel praesentium non pariatur. Omnis ut magni voluptas et. Ducimus harum sed beatae labore.",
+      created: "2009-09-25T07:04:00.000Z",
+      modified: "2021-12-09T12:01:33.000Z",
+      modifiedBy: "",
+      createdBy: "",
+      templateInfo: {
+        "@id": "/v2/templates/00HH42EEHC05QT14VQ041R1KEY",
+        options: [],
+      },
+      theme: "",
+      onPlaylists: ["/v2/playlists/00S7ZQK8Y90R351YES1DJN0RKR"],
+      duration: 11240,
+      published: {
+        from: "2020-12-24T06:17:03.000Z",
+        to: null,
+      },
+      media: ["/v2/media/000NDA6BBM1A071RNN0SP40XNJ"],
+    },
+    {
+      "@id": "/v2/slides/00J96KHGSC1HPH0VHW1FSX0CVD",
+      title: "Assumenda quaerat sint nihil perspiciatis.",
+      description:
+        "Eaque quaerat voluptas vitae rerum numquam dolore explicabo. Quas error voluptates quo est vel. Odit cum illum qui aut iure ipsum officiis.",
+      created: "1989-11-29T16:39:50.000Z",
+      modified: "2021-12-09T12:01:33.000Z",
+      modifiedBy: "",
+      createdBy: "",
+      templateInfo: {
+        "@id": "/v2/templates/001R8FR6VC10G51B200TK60QP3",
+        options: [],
+      },
+      theme: "/v2/themes/01FPFH3WX93S4575W6Q9T8K0YN",
+      onPlaylists: [],
+      duration: 81995,
+      published: {
+        from: "2021-11-26T14:53:28.000Z",
+        to: null,
+      },
+      media: [
+        "/v2/media/009H64MSPN1HEH0DTV2DEV085B",
+        "/v2/media/00H1A2WZ1A06BK0W781ETS1EXZ",
+        "/v2/media/01BBCV60M903YC07RP09VN1JBZ",
+      ],
+    },
+    {
+      "@id": "/v2/slides/00FTMVQRPS18QD0MBY0RQ30FP0",
+      title: "Atque nulla explicabo beatae.",
+      description:
+        "Repudiandae sapiente voluptatibus dolores voluptatem. Recusandae quia iste voluptates tempora fuga aliquam aut hic.",
+      created: "1987-03-29T10:52:22.000Z",
+      modified: "2021-12-09T12:01:33.000Z",
+      modifiedBy: "",
+      createdBy: "",
+      templateInfo: {
+        "@id": "/v2/templates/01F8MJT0AV24D602FN131008YY",
+        options: [],
+      },
+      theme: "/v2/themes/01FPFH3WX93S4575W6Q9T8K0Y9",
+      onPlaylists: [],
+      duration: 32646,
+      published: {
+        from: "2021-09-04T10:08:04.000Z",
+        to: "2021-10-27T01:46:23.000Z",
+      },
+      media: [
+        "/v2/media/000NDA6BBM1A071RNN0SP40XNJ",
+        "/v2/media/003M4E1AQG052J0HPM1X670P4F",
+        "/v2/media/009RF9AGF80QP60ECM0CY01A1Y",
+        "/v2/media/00QWNXTD5X0VSF1EWQ1H7H0YS2",
+      ],
+    },
+    {
+      "@id": "/v2/slides/008J6BKZMA0YHM04S813NF0DZV",
+      title: "Atque quis aut qui veritatis.",
+      description:
+        "Dolore expedita labore praesentium beatae. Labore maiores est voluptates cupiditate praesentium. Maiores est dolorem iusto.",
+      created: "1979-05-01T14:59:35.000Z",
+      modified: "2021-12-09T12:01:33.000Z",
+      modifiedBy: "",
+      createdBy: "",
+      templateInfo: {
+        "@id": "/v2/templates/011763DN4P0FRS1T6B09180KCW",
+        options: [],
+      },
+      theme: "",
+      onPlaylists: ["/v2/playlists/01EDTHWKBM0TR3035E0Z631B8E"],
+      duration: 86942,
+      published: {
+        from: "2020-12-26T12:37:10.000Z",
+        to: null,
+      },
+      media: [
+        "/v2/media/00JBVT18N31VBA0S3805CN03MC",
+        "/v2/media/016SXZGDJR0T321KYJ19B91N1Z",
+        "/v2/media/01795VPQZX0GNM00T11JJ71BS8",
+      ],
+    },
+    {
+      "@id": "/v2/slides/0118GD1EXJ1CSM1BHW0P7K1KQG",
+      title: "Aut odio illum et.",
+      description:
+        "Voluptatibus dicta dolor suscipit eos. Voluptate atque quia quo blanditiis sint est odio. Voluptatum expedita ipsa ut et placeat laborum. Modi molestias quibusdam sunt ratione nulla sit.",
+      created: "2006-03-22T07:17:31.000Z",
+      modified: "2021-12-09T12:01:33.000Z",
+      modifiedBy: "",
+      createdBy: "",
+      templateInfo: {
+        "@id": "/v2/templates/000AT4BWMM1GJZ17WE1M9M0RFB",
+        options: [],
+      },
+      theme: "/v2/themes/01FPFH3WX93S4575W6Q9T8K0YB",
+      onPlaylists: [],
+      duration: 104864,
+      published: {
+        from: "2021-03-27T10:02:38.000Z",
+        to: null,
+      },
+      media: [
+        "/v2/media/002B3H5Z2K10M211AK1ERC02PM",
+        "/v2/media/00XA10EH8H0YYY1H4B1CC40SSA",
+      ],
+    },
+    {
+      "@id": "/v2/slides/0151GSJGTE1JSE032C14JR0CQ2",
+      title: "Aut qui placeat consequatur aut.",
+      description:
+        "Officiis sit adipisci exercitationem ut vero nihil iste voluptas. Temporibus aut saepe rerum sint. Soluta perferendis aliquid voluptatem eos.",
+      created: "2010-05-04T04:35:53.000Z",
+      modified: "2021-12-09T12:01:33.000Z",
+      modifiedBy: "",
+      createdBy: "",
+      templateInfo: {
+        "@id": "/v2/templates/01F8MJT0AV24D602FN131008YY",
+        options: [],
+      },
+      theme: "",
+      onPlaylists: [],
+      duration: 28141,
+      published: {
+        from: "2020-12-14T13:20:43.000Z",
+        to: "2021-02-11T10:45:45.000Z",
+      },
+      media: [
+        "/v2/media/00KXYB7Z291JXC1SY30G161HQD",
+        "/v2/media/01ENQN7X2F0ANK1BNA0J3513JX",
+      ],
+    },
+  ],
+  "hydra:totalItems": 100,
+};
+
+const themesJson = {
+  "@context": "/contexts/Theme",
+  "@id": "/v2/themes",
+  "@type": "hydra:Collection",
+  "hydra:member": [
+    {
+      "@type": "Theme",
+      "@id": "/v2/themes/01FTNTE788816N6YCW9NVM2JQB",
+      title: "Consequatur quisquam recusandae asperiores accusamus.",
+      description:
+        "Occaecati debitis et saepe eum sint dolorem. Enim ipsum inventore sed libero et velit qui suscipit. Deserunt laudantium quibusdam enim nostrum soluta qui ipsam non.",
+      onSlides: [],
+      created: "2022-01-30T15:42:42+01:00",
+      modified: "2022-01-30T15:42:42+01:00",
+      modifiedBy: "",
+      createdBy: "",
+      css: "",
+    },
+    {
+      "@type": "Theme",
+      "@id": "/v2/themes/01FTNTE788816N6YCW9NVM2JQC",
+      title: "Sit vitae voluptas sint non.",
+      description:
+        "Optio quos qui illo error. Laborum vero a officia id corporis. Saepe provident esse hic eligendi. Culpa ut ab voluptas sed a.",
+      onSlides: [
+        {
+          "@type": "Slide",
+          "@id": "/v2/slides/007ZR8C9811R741WAP1KHK0T09",
+          title: "Mollitia iure culpa exercitationem.",
+          description:
+            "Facilis nihil minus vel eum. Ut corrupti dicta quo modi temporibus est.",
+          created: "1978-09-14T10:51:02+01:00",
+          modified: "2022-01-30T15:42:42+01:00",
+          modifiedBy: "",
+          createdBy: "",
+          templateInfo: {
+            "@id": "/v2/templates/017BG9P0E0103F0TFS17FM016M",
+            options: [],
+          },
+          theme: "/v2/themes/01FTNTE788816N6YCW9NVM2JQC",
+          onPlaylists: [],
+          duration: 92789,
+          published: {
+            from: "2021-08-04T05:24:19+02:00",
+            to: "2021-11-14T19:59:53+01:00",
+          },
+          media: ["/v2/media/00MTYFFTN30XNZ0F350YM31TN5"],
+          content: [],
+          feed: null,
+        },
+        {
+          "@type": "Slide",
+          "@id": "/v2/slides/00GYC65JP01DEG1F6H0R5Q0JBA",
+          title: "Voluptatem recusandae hic.",
+          description:
+            "Nulla occaecati praesentium quia magni ipsum dolor. Et aliquid natus molestiae ut quis. Ad voluptatum qui consequatur deleniti labore est. Voluptas hic veritatis quidem molestias qui.",
+          created: "1988-06-15T11:26:36+02:00",
+          modified: "2022-01-30T15:42:42+01:00",
+          modifiedBy: "",
+          createdBy: "",
+          templateInfo: {
+            "@id": "/v2/templates/002BAP34VD1EHG0E4J0D2Y00JW",
+            options: [],
+          },
+          theme: "/v2/themes/01FTNTE788816N6YCW9NVM2JQC",
+          onPlaylists: [],
+          duration: 78938,
+          published: {
+            from: "2021-10-09T12:14:12+02:00",
+            to: "2021-12-24T16:18:08+01:00",
+          },
+          media: [
+            "/v2/media/00YB5658GH0TAE1A1N0XBB0YR7",
+            "/v2/media/01CVTNA9Y917EX09MY0FNX0GKA",
+            "/v2/media/01E4S5SPXR19MP1KGY16TD1XG2",
+          ],
+          content: [],
+          feed: null,
+        },
+        {
+          "@type": "Slide",
+          "@id": "/v2/slides/00V28394SD1WBY1BPE1STN13MD",
+          title: "Reprehenderit neque nam mollitia quia.",
+          description:
+            "Omnis aliquam ea architecto dignissimos. Harum provident asperiores neque consequatur sit sed. Quasi ipsa illum et qui deleniti quo.",
+          created: "1999-06-23T10:05:00+02:00",
+          modified: "2022-01-30T15:42:42+01:00",
+          modifiedBy: "",
+          createdBy: "",
+          templateInfo: {
+            "@id": "/v2/templates/01FGC8EXSE1KCC1PTR0NHB0H3R",
+            options: [],
+          },
+          theme: "/v2/themes/01FTNTE788816N6YCW9NVM2JQC",
+          onPlaylists: [],
+          duration: 69299,
+          published: {
+            from: "2021-06-09T03:25:34+02:00",
+            to: "2021-11-05T02:30:21+01:00",
+          },
+          media: ["/v2/media/0041NS3DFY1EMS0XGQ025B0425"],
+          content: [],
+          feed: null,
+        },
+      ],
+      created: "2022-01-30T15:42:42+01:00",
+      modified: "2022-01-30T15:42:42+01:00",
+      modifiedBy: "",
+      createdBy: "",
+      css: "",
+    },
+    {
+      "@type": "Theme",
+      "@id": "/v2/themes/01FTNTE788816N6YCW9NVM2JQD",
+      title: "Enim ex eveniet facere.",
+      description:
+        "Delectus aut nam et eum. Fugit repellendus illo veritatis. Ex esse veritatis voluptate vel possimus. Aut incidunt sunt cumque asperiores incidunt iure sequi.",
+      onSlides: [
+        {
+          "@type": "Slide",
+          "@id": "/v2/slides/003VYYZPPN1MQ61MEM1TE10G2J",
+          title: "Quos ducimus culpa consequuntur nulla aliquid.",
+          description:
+            "At quia quia voluptatibus eius. Delectus quia consequuntur aut nihil. Impedit sit aut dolorum aut dolore. Dolore beatae ipsa voluptas.",
+          created: "1974-03-21T14:49:33+01:00",
+          modified: "2022-01-30T15:42:42+01:00",
+          modifiedBy: "",
+          createdBy: "",
+          templateInfo: {
+            "@id": "/v2/templates/01FGC8EXSE1KCC1PTR0NHB0H3R",
+            options: [],
+          },
+          theme: "/v2/themes/01FTNTE788816N6YCW9NVM2JQD",
+          onPlaylists: ["/v2/playlists/00XVQEW1EV0N3K0JZQ0TYS0T1G"],
+          duration: 77194,
+          published: {
+            from: "2021-07-02T08:33:02+02:00",
+            to: "2022-03-12T20:52:07+01:00",
+          },
+          media: [
+            "/v2/media/00GEQ02WW10SZ21F9G1MAZ0KR8",
+            "/v2/media/00MTYFFTN30XNZ0F350YM31TN5",
+            "/v2/media/00SSYSBFHR16PM09MQ0B0K0202",
+            "/v2/media/00X6A9GBZM0EHF05AA0B350D8E",
+          ],
+          content: [],
+          feed: null,
+        },
+        {
+          "@type": "Slide",
+          "@id": "/v2/slides/00EBKSK8ZZ0Y301VFR0WGQ05F4",
+          title: "Maiores repudiandae quibusdam et rerum.",
+          description:
+            "At totam ut animi nisi ut ut qui. Aspernatur omnis quod temporibus non quo numquam. Dignissimos non eius numquam neque. Numquam modi tempora minus ad aut aut sit.",
+          created: "1985-08-21T22:37:57+02:00",
+          modified: "2022-01-30T15:42:42+01:00",
+          modifiedBy: "",
+          createdBy: "",
+          templateInfo: {
+            "@id": "/v2/templates/0044JYNRTJ1KD0128318R80B3Q",
+            options: [],
+          },
+          theme: "/v2/themes/01FTNTE788816N6YCW9NVM2JQD",
+          onPlaylists: [],
+          duration: 66516,
+          published: {
+            from: "2021-08-30T14:57:29+02:00",
+            to: "2021-10-24T12:24:48+02:00",
+          },
+          media: ["/v2/media/00031XCV6Z1W860V5R0SXB0D6R"],
+          content: [],
+          feed: null,
+        },
+        {
+          "@type": "Slide",
+          "@id": "/v2/slides/00VRS9JQ7C1BZZ1NBA1XE50E19",
+          title: "Doloremque cum aliquam quis sint.",
+          description:
+            "Est quos beatae voluptatem optio et sit. Culpa fugiat quam et quisquam error a. Aut molestias quaerat quia aut non ipsum autem. Sunt aspernatur eos dolores quas alias. Culpa aut maiores consectetur.",
+          created: "2000-03-29T12:07:31+02:00",
+          modified: "2022-01-30T15:42:42+01:00",
+          modifiedBy: "",
+          createdBy: "",
+          templateInfo: {
+            "@id": "/v2/templates/01FGC8EXSE1KCC1PTR0NHB0H3R",
+            options: [],
+          },
+          theme: "/v2/themes/01FTNTE788816N6YCW9NVM2JQD",
+          onPlaylists: [],
+          duration: 28295,
+          published: {
+            from: "2022-03-18T13:47:21+01:00",
+            to: "2022-03-19T12:34:17+01:00",
+          },
+          media: [
+            "/v2/media/003NVKRN4E183T0C431JF7036P",
+            "/v2/media/00C07KS3R00PEV24RF09870SH9",
+            "/v2/media/0170X462SF1P3205JP1R6K0553",
+          ],
+          content: [],
+          feed: null,
+        },
+        {
+          "@type": "Slide",
+          "@id": "/v2/slides/00X2XD9Y011VB31K2T10JW0NR1",
+          title: "Eveniet repellendus et autem repellat.",
+          description:
+            "Rerum praesentium quo sequi. Accusamus fugiat voluptatem est quam. Esse voluptatem quia fugiat nisi delectus omnis.",
+          created: "2001-09-04T01:28:51+02:00",
+          modified: "2022-01-30T15:42:42+01:00",
+          modifiedBy: "",
+          createdBy: "",
+          templateInfo: {
+            "@id": "/v2/templates/017BG9P0E0103F0TFS17FM016M",
+            options: [],
+          },
+          theme: "/v2/themes/01FTNTE788816N6YCW9NVM2JQD",
+          onPlaylists: ["/v2/playlists/007TM6JDGF1ECH07J10ZHY0S7P"],
+          duration: 99209,
+          published: {
+            from: "2021-11-11T07:58:12+01:00",
+            to: "2021-11-13T19:36:00+01:00",
+          },
+          media: [
+            "/v2/media/00J8PGYF1N12T60T200QN50KKJ",
+            "/v2/media/014S7FGP500Z8P18ZW12631TCP",
+            "/v2/media/019PTTMBQB0Z2D05VQ0HVC1M5M",
+          ],
+          content: [],
+          feed: null,
+        },
+        {
+          "@type": "Slide",
+          "@id": "/v2/slides/011KV2WHQS0NDK1Q1Q01GY0XPS",
+          title: "Harum ducimus reiciendis.",
+          description:
+            "Est aut quis omnis. Cumque id officiis molestias accusamus est molestias. Nulla qui aut quo sunt et.",
+          created: "2006-08-10T03:26:54+02:00",
+          modified: "2022-01-30T15:42:42+01:00",
+          modifiedBy: "",
+          createdBy: "",
+          templateInfo: {
+            "@id": "/v2/templates/01FGC8EXSE1KCC1PTR0NHB0H3R",
+            options: [],
+          },
+          theme: "/v2/themes/01FTNTE788816N6YCW9NVM2JQD",
+          onPlaylists: [],
+          duration: 7509,
+          published: {
+            from: "2021-07-22T07:48:05+02:00",
+            to: "2021-09-03T23:11:43+02:00",
+          },
+          media: [
+            "/v2/media/008ARWMTYJ0SX810490JQQ0DAK",
+            "/v2/media/00E98GAQXH1Y2C1G131MVH0YWZ",
+            "/v2/media/00MTYFFTN30XNZ0F350YM31TN5",
+            "/v2/media/0142VZYZ7H0XHE1XJB1M730VGS",
+            "/v2/media/015H8MEPVH13BE15A11MHJ0KAM",
+          ],
+          content: [],
+          feed: null,
+        },
+      ],
+      created: "2022-01-30T15:42:42+01:00",
+      modified: "2022-01-30T15:42:42+01:00",
+      modifiedBy: "",
+      createdBy: "",
+      css: "",
+    },
+  ],
+  "hydra:totalItems": 20,
+};
+
+const themesSingleJson = {
+  "@type": "Theme",
+  "@id": "/v2/themes/01FTNTE788816N6YCW9NVM2JQN",
+  title: "Hic minus et omnis porro.",
+  description:
+    "Odit quia nisi accusantium natus. Ut explicabo corporis eligendi ut. Sapiente ut qui quidem explicabo optio amet velit aut. Iure sed alias asperiores perspiciatis deserunt omnis inventore mollitia.",
+  onSlides: [],
+  created: "2022-01-30T15:42:42+01:00",
+  modified: "2022-01-30T15:42:42+01:00",
+  modifiedBy: "",
+  createdBy: "",
+  css: "",
+};
+
 export {
   tokenAdminJson,
   tokenTenantsJson,
@@ -1027,4 +1663,8 @@ export {
   screenGroupsListJson,
   screenGroupsSingleJson,
   screensListJson,
+  templatesListJson,
+  slidesListJson,
+  themesJson,
+  themesSingleJson,
 };
